+++ rl_evaluate.py
@@ -3,7 +3,7 @@ from ray.rllib.models import ModelCatalog
 from rl_agent.rl_env import TiramisuRlEnv
 from ray.rllib.algorithms.ppo import PPO, PPOConfig
 from ray.rllib.algorithms.algorithm_config import AlgorithmConfig
-from rllib_ray_utils.dataset_actor import DatasetActor, DatasetFormat
+from rllib_ray_utils.dataset_actor.dataset_actor import DatasetActor, DatasetFormat
 from config.config import Config
 from rl_agent.rl_policy_nn import PolicyNN
