+++ rllib_ray_utils/dataset_actor/services/hybrid_data_service.py
@@ -1,15 +1,14 @@
 import pickle
 import random
-from typing import Tuple
-
+import os
 import numpy as np
 from env_api.core.models.tiramisu_program import TiramisuProgram
-from env_api.data.services.base_data_service import (
+from rllib_ray_utils.dataset_actor.services.base_data_service import (
     BaseDataService,
 )
 
 
-class PickleDataService(BaseDataService):
+class HybridDataService(BaseDataService):
     def __init__(
         self,
         dataset_path: str,
@@ -30,15 +29,12 @@ class PickleDataService(BaseDataService):
         self.cpps = {}
 
         print(
-            f"reading dataset in full pkl format: dataset pkl from {self.dataset_path} and cpps pkl from {self.cpps_path}"
+            f"reading dataset in Hybrid format: dataset pkl from {self.dataset_path} and cpps from {self.cpps_path}"
         )
 
         with open(self.dataset_path, "rb") as f:
             self.dataset = pickle.load(f)
-            self.function_names = list(self.dataset.keys())
-
-        with open(self.cpps_path, "rb") as f:
-            self.cpps = pickle.load(f)
+            self.function_names = os.listdir(self.cpps_path)
 
         # Shuffle the dataset (can be used with random sampling turned off to get a random order)
         if self.shuffle:
@@ -63,5 +59,7 @@ class PickleDataService(BaseDataService):
         print(
             f"Selected function with index: {self.current_function_index}, name: {function_name}"
         )
+        file_name = function_name + "_generator.cpp"
+        file_path = self.cpps_path + function_name + "/" + file_name
 
-        return TiramisuProgram.from_dict(function_name, self.dataset[function_name], self.cpps[function_name])
+        return TiramisuProgram.from_dict(function_name, self.dataset[function_name])
